requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

discord.py==2.3.0
python-dotenv==1.0.0
python-weather==0.4.3
bruhcolor==0.0.56
matplotlib==3.6.3
youtube-dl==2021.2.10
prettytable==3.8.0
wcwidth==0.2.6
pynacl==1.5.0
cffi==1.15.1
pycparser==2.21
discordwebhook
